Fixed bug in enumeration, made enumeration functions public
The second parameter of SetupDiEnumDeviceInterfaces was previously
hardcoded to 0. This caused the function to give a list of all
interfaces on all devices. As the enumeration function is already
looping through all devices via SetupDiEnumDeviceInfo, this caused each
of n interfaces to be enumerated n times, resulting in n^2 results. The
workaround was to check every interface to verify it was unique, which
can be seen on line 68. I removed the workaround because it is no longer
needed.