Fix Native Methods to Utilize Byte Array
The methods are incorrectly passing a single byte. See the following definitions (all identical).
https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/hidsdi/nf-hidsdi-hidd_getmanufacturerstring
https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/hidsdi/nf-hidsdi-hidd_getproductstring
https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/hidsdi/nf-hidsdi-hidd_getserialnumberstring
All of these take pvoid (void pointer), so we pass the byte array, the interface will get the array pointer, and fill it accordingly. If we pass a single byte, c# won't map the array correctly in the return data.